requirements.txt
fastapi
uvicorn
lark-oapi
langgraph
langchain-openai
langchain-core
python-dotenv
feedparser
apscheduler
typing-extensions
